app/requirements.txt
asgiref==3.4.1
autopep8==1.5.7
black==21.8b0
click==8.0.1
colorama==0.4.4
deta==1.0.0
fastapi==0.68.1
h11==0.12.0
httptools==0.2.0
mypy-extensions==0.4.3
orjson==3.6.3
pathspec==0.9.0
platformdirs==2.3.0
pycodestyle==2.7.0
pydantic==1.8.2
python-dotenv==0.19.0
python-multipart==0.0.5
PyYAML==5.4.1
regex==2021.8.28
shortuuid==1.0.1
six==1.16.0
starlette==0.14.2
toml==0.10.2
tomli==1.2.1
typing-extensions==3.10.0.2
uvicorn==0.15.0
watchgod==0.7
websockets==9.1
